fixed a couple of warnings

+++ src/AssetLoader/TextureLoader.cpp
@@ -12,9 +12,9 @@ char* TextureLoader::LoadTexture(const std::string& filename, TextureInfo& info)
 	info.Height = texture.extent().y;
 	info.Layers = texture.extent().z;
 	info.Format = texture.format();
-	info.MipCount = texture.levels();
-	info.LinearSize = texture.size();
-	info.BPP = (texture.size(0) * 8) / (texture.extent(0).x * texture.extent(0).y); //bits for whole mip 0 / number of pixels
+	info.MipCount = uint32_t(texture.levels());
+	info.LinearSize = uint32_t(texture.size());
+	info.BPP = uint32_t((texture.size(0) * 8) / (texture.extent(0).x * texture.extent(0).y)); //bits for whole mip 0 / number of pixels
 	info.Data = malloc(texture.size());
 	memcpy(info.Data, texture.data(), texture.size());
 	return nullptr;
