+++ src/Core/components/TransformComponent.h
@@ -19,9 +19,9 @@ static glm::quat EulertoQuaternion(float pitch, float roll, float yaw)
 	double t4 = std::cos(pitch * 0.5);
 	double t5 = std::sin(pitch * 0.5);
 
-	q.w = t0 * t2 * t4 + t1 * t3 * t5;
-	q.x = t0 * t3 * t4 - t1 * t2 * t5;
-	q.y = t0 * t2 * t5 + t1 * t3 * t4;
-	q.z = t1 * t2 * t4 - t0 * t3 * t5;
+	q.w = float(t0 * t2 * t4 + t1 * t3 * t5);
+	q.x = float(t0 * t3 * t4 - t1 * t2 * t5);
+	q.y = float(t0 * t2 * t5 + t1 * t3 * t4);
+	q.z = float(t1 * t2 * t4 - t0 * t3 * t5);
 	return q;
 }
