+++ src/Graphics/Raymarch.cpp
@@ -131,8 +131,8 @@ void Raymarcher::Render(VulkanCommandBuffer& cmdBuffer, uint32_t frameIndex, vk:
 	cmdBuffer.bindPipeline(vk::PipelineBindPoint::eCompute, m_Pipeline.GetPipeline());
 	cmdBuffer.bindDescriptorSets(vk::PipelineBindPoint::eCompute, m_Pipeline.GetPipelineLayout(), 0, m_DescSets[frameIndex], nullptr);
 	cmdBuffer.bindDescriptorSets(vk::PipelineBindPoint::eCompute, m_Pipeline.GetPipelineLayout(), 1, ibl, nullptr);
-	const int gs = 32;
-	uint32_t x = (screenSize.x + gs - 1) / gs;
-	uint32_t y = (screenSize.y + gs - 1) / gs;
+	const uint32_t gs = 32;
+	uint32_t x = ((uint32_t)screenSize.x + gs - 1) / gs;
+	uint32_t y = ((uint32_t)screenSize.y + gs - 1) / gs;
 	cmdBuffer.dispatch(x, y, 1);
 }
